another overdue reorganizing of the private and public pages + layouts
