+++ client/public/static/styles/layer_cake/style.css
@@ -231,7 +231,6 @@ ul.thin li {
 }
 
 #logo {
-  background: url("./images/logo-new.png") no-repeat center center;
   height: 100px;
   margin: 0 auto;
   width: 650px;
