+++ client/src/components/pages/private/layout/PrivateFooter.js
