+++ client/src/components/pages/private/layout/sidebar/FeaturedAlbum.js
