+++ client/src/components/pages/private/layout/sidebar/VanityAlbum.js
