--- client/src/components/routing/PrivateRoute.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Navigate, useLocation } from 'react-router-dom';
-import { connect } from 'react-redux';
-
-const PrivateRoute = ({ isAuthenticated, children }) => {
-  const location = useLocation();
-
-  return (
-    isAuthenticated ? (
-      children
-    ) : (
-      <Navigate to="/login" state={{ from: location }} replace />
-    )
-  );
-};
-
-const mapStateToProps = state => ({
-  isAuthenticated: state.auth.isAuthenticated
-});
-
-export default connect(mapStateToProps)(PrivateRoute);
